API

Model Interface

CalibrateAtmos.get_config — Function

get_config(member, iteration, experiment_id::AbstractString)
get_config(member, iteration, config_dict::AbstractDict)

Returns an AtmosConfig object for the given member and iteration. If given an experiment id string, it will load the config from the corresponding YAML file. Turns off default diagnostics and sets the TOML parameter file to the member's path. This assumes that the config dictionary has output_dir and restart_file keys.

CalibrateAtmos.run_forward_model — Function

run_forward_model(config::AbstractDict)

Runs the atmosphere model with the given an AtmosConfig object. Currently only has basic error handling.

CalibrateAtmos.get_forward_model — Function

get_forward_model(experiment_id::Val)

Returns the custom physical model objet for the given experiment id. An error is thrown if the experiment id is not recognized.

CalibrateAtmos.observation_map — Function

observation_map(physical_model::AbstractPhysicalModel, iteration)

Returns the observation for the given case id Value and iteration.

NB: ensure that the model output is sufficiently sensitive to the input parameters.

EnsembleKalmanProcesses Interface

CalibrateAtmos.initialize — Function

initialize(
    experiment_id;
    config = YAML.load_file("experiments/$experiment_id/ekp_config.yml"),
    Γ = JLD2.load(config["truth_noise"]),
    y = JLD2.load(config["truth_data"]),
    rng_seed = 1234,
)

Initializes the EKP object and the model ensemble.

Takes in

experiment_id: the name of the experiment, which corresponds to the name of the subfolder in experiments/

config: a dictionary of configuration values

CalibrateAtmos.update_ensemble — Function

update_ensemble(
    experiment_id,
    iteration;
    config = YAML.load_file("experiments/$experiment_id/ekp_config.yml"),
)

Updates the EKI object and saves parameters for the next iteration. Assumes that the observation map has been run and saved in the current iteration folder.